# apps/api/internal/store/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
	pgvector_go "github.com/pgvector/pgvector-go"
)

type ApiKey struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	Name       string
	KeyHash    []byte
	Prefix     string
	CreatedAt  pgtype.Timestamptz
	LastUsedAt pgtype.Timestamptz
	RevokedAt  pgtype.Timestamptz
}

type Asset struct {
	ID               uuid.UUID
	UserID           uuid.UUID
	ItemID           pgtype.UUID
	ContentType      string
	ByteSize         int64
	OriginalFilename string
	CreatedAt        pgtype.Timestamptz
}

type DeviceLink struct {
	CodeHash   []byte
	UserID     uuid.UUID
	DeviceHint string
	CreatedAt  pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
	ClaimedAt  pgtype.Timestamptz
}

type Feed struct {
	ID                  uuid.UUID
	UserID              uuid.UUID
	Url                 string
	Title               string
	SiteUrl             string
	LastPolledAt        pgtype.Timestamptz
	LastStatus          string
	CreatedAt           pgtype.Timestamptz
	Etag                string
	LastModified        string
	NextPollAt          pgtype.Timestamptz
	PollIntervalMinutes int32
}

type Highlight struct {
	ID           uuid.UUID
	UserID       uuid.UUID
	SourceItemID uuid.UUID
	QuoteItemID  uuid.UUID
	Exact        string
	Prefix       string
	Suffix       string
	OffsetHint   int32
	CreatedAt    pgtype.Timestamptz
}

type Item struct {
	ID             uuid.UUID
	UserID         uuid.UUID
	Url            string
	Title          string
	Body           string
	LeadImageUrl   string
	Summary        string
	Tags           []string
	CardType       string
	Status         string
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
	Palette        []string
	UserTags       []string
	PinnedAt       pgtype.Timestamptz
	LastDriftedAt  pgtype.Timestamptz
	SearchTsv      interface{}
	PageCount      pgtype.Int4
	FeedID         pgtype.UUID
	KeptAt         pgtype.Timestamptz
	TaggedLocation string
	UrlHost        pgtype.Text
	BodyMarkdown   pgtype.Text
}

type ItemEmbedding struct {
	ItemID    uuid.UUID
	UserID    uuid.UUID
	Embedding pgvector_go.Vector
	CreatedAt pgtype.Timestamptz
}

type ItemPlace struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	ItemID    uuid.UUID
	Name      string
	Hint      string
	Address   string
	Lat       pgtype.Float8
	Lng       pgtype.Float8
	Source    string
	CreatedAt pgtype.Timestamptz
}

type JevDecision struct {
	ID            int64
	UserID        uuid.UUID
	ItemID        pgtype.UUID
	Surface       string
	Model         string
	QuestionsV    string
	Answers       []byte
	Action        string
	UserVerdict   pgtype.Text
	LatencyMs     pgtype.Int4
	InputTokens   pgtype.Int4
	CreatedAt     pgtype.Timestamptz
	AppliedTags   []string
	SuggestedTags []string
	DismissedTags []string
}

type Lense struct {
	ID             uuid.UUID
	UserID         uuid.UUID
	Name           string
	Rule           []byte
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
	DigestSchedule string
	LastDigestAt   pgtype.Timestamptz
}

type Link struct {
	UserID    uuid.UUID
	AItem     uuid.UUID
	BItem     uuid.UUID
	CreatedAt pgtype.Timestamptz
}

type Notification struct {
	ID           uuid.UUID
	UserID       uuid.UUID
	Category     string
	DedupeKey    string
	Title        string
	Body         string
	Data         []byte
	CreatedAt    pgtype.Timestamptz
	DeliverAfter pgtype.Timestamptz
	SentAt       pgtype.Timestamptz
	Attempts     int32
	LastError    string
}

type NotificationDelivery struct {
	ID             uuid.UUID
	UserID         uuid.UUID
	NotificationID uuid.UUID
	Channel        string
	Token          string
	TicketID       string
	SentAt         pgtype.Timestamptz
	Ok             bool
	Error          string
}

type PushDevice struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	ApiKeyID   pgtype.UUID
	Token      string
	Platform   string
	CreatedAt  pgtype.Timestamptz
	LastSeenAt pgtype.Timestamptz
	FailedAt   pgtype.Timestamptz
}

type User struct {
	ID          uuid.UUID
	CreatedAt   pgtype.Timestamptz
	ClerkUserID pgtype.Text
	Email       string
}

type UserSetting struct {
	UserID    uuid.UUID
	Key       string
	Value     string
	UpdatedAt pgtype.Timestamptz
}