# openapi.yaml
openapi: 3.0.3
info: { title: Openmind API, version: 0.0.1 }
security:
  - bearerAuth: []
paths:
...
  /items/{id}:
    get:
      operationId: getItem
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: item detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemDetail" }
        "404":
          description: not found
    patch:
      operationId: patchItem
      description: "Edit an item. Currently supports replacing the full user-tags list (userTags). Tags are canonicalised server-side (trimmed, lowercased, deduped, capped)."
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateItemRequest" }
      responses:
        "200":
          description: updated item detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemDetail" }
        "400": { description: bad request }
        "404": { description: not found }
    delete:
      operationId: deleteItem
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: deleted }
        "404": { description: not found }
  /items/{id}/jev-verdict:
    post:
      operationId: postJevVerdict
      summary: Record a user reaction to a Jev capture suggestion or auto-applied tag
      description: >
        Backfills jev_decisions.user_verdict when the user accepts a mid-confidence
        suggestion chip, dismisses one, undoes an auto-applied tag, or confirms
        an auto-apply. Accept also adds the tag to userTags; undo removes it.
        Idempotent for the same tag+action. No-op (404) when the item has no
        capture decision.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JevVerdictRequest" }
      responses:
        "200":
          description: updated item detail (includes refreshed jevSuggestions)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemDetail" }
        "400": { description: bad request }
        "404": { description: not found }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  schemas:
...
    Item:
      type: object
      required: [id, url, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        title: { type: string }
        summary: { type: string }
        leadImageUrl: { type: string }
        tags: { type: array, items: { type: string } }
        userTags: { type: array, items: { type: string } }
        palette: { type: array, items: { type: string } }
        cardType: { type: string, enum: [article, product, book, recipe, video, tweet, image, note, quote, repo] }
        status: { type: string, enum: [pending, enriched, failed] }
        createdAt: { type: string, format: date-time }
        pinnedAt: { type: string, format: date-time, nullable: true, description: "When the item was pinned to the Desk; null if not pinned." }
        pageCount: { type: integer, nullable: true, description: "Page count when the item's original is a stored PDF; null otherwise." }
        feedId: { type: string, format: uuid, nullable: true, description: "The feed this item originated from; null if not feed-sourced." }
        keptAt: { type: string, format: date-time, nullable: true, description: "When the item was kept in the library from its feed; null if not kept." }
...
    ItemDetail:
      allOf:
        - $ref: "#/components/schemas/Item"
        - type: object
          required: [body]
          properties:
            body: { type: string }
            jevSuggestions:
              $ref: "#/components/schemas/JevSuggestions"
              description: >
                Phase 2 capture chips for this item when AI-assisted organisation
                produced applied/suggested tags. Absent when there is no capture
                decision or nothing left to show.
    JevSuggestions:
      type: object
      required: [action, suggestedTags, appliedTags]
      properties:
        action:
          type: string
          description: "Row-level capture action: applied | suggested | skipped | shadow"
        suggestedTags:
          type: array
          items: { type: string }
          description: Mid-confidence tags not yet accepted, dismissed, or present on the item
        appliedTags:
          type: array
          items: { type: string }
          description: Auto-applied tags still on the item (one-tap undo)
        userVerdict:
          type: string
          nullable: true
          description: kept | changed | removed when the user has reacted
    JevVerdictRequest:
      type: object
      required: [action, tag]
      properties:
        action:
          type: string
          enum: [accept, dismiss, undo, keep]
          description: >
            accept adds a suggested tag (verdict kept); dismiss hides a suggestion
            (verdict removed); undo removes an auto-applied tag (verdict removed);
            keep confirms an auto-apply without changing tags (verdict kept).
        tag:
          type: string
          description: The tag the user reacted to
...
    UpdateItemRequest:
      type: object
      description: "Fields to update on an item. Both userTags and pinned are optional; omit both for a no-op edit that is rejected as a bad request."
      properties:
        userTags: { type: array, items: { type: string }, description: "Full replacement user-tags list. An empty array clears all user tags." }
        pinned: { type: boolean, description: "Pin (true) or unpin (false) the item on the Desk. Pinning sets pinnedAt to now; unpinning clears it." }
        kept: { type: boolean, description: "Keep (true) the item in the library independent of its feed. Sets keptAt to now." }
...